CHANGE: 리뷰목록API, 코드리뷰 반영
