+++ src/controllers/index.js
@@ -1,7 +1,11 @@
 const usersController = require("./users.controller");
 const productsController = require("./products.controller");
+const likesController = require("./likes.controller");
+const reviewsController = require("./reviews.controller");
 
 module.exports = {
   usersController,
   productsController,
+  likesController,
+  reviewsController,
 };
