+++ src/controllers/products.controller.js
@@ -1,47 +1,29 @@
 const { productsService } = require("../services");
 
-const createLike = async (req, res) => {
+const getProduct = async (req, res) => {
   try {
-    const { productId } = req.params;
-    // const { userId } = req;
-    const userId = req.get("userId");
-    await productsService.createLike(userId, productId);
-    res.status(201).json({ message: "likeCreated" });
-  } catch (error) {
-    console.log(error);
-    res.status(error.status).json({ message: error.message });
-  }
-};
-
-const deleteLike = async (req, res) => {
-  try {
-    const { productId } = req.params;
-    // const { userId } = req;
-    const userId = req.get("userId");
-    await productsService.deleteLike(userId, productId);
-    res.status(204).json({ message: "likeDeleted" });
-  } catch (error) {
-    console.log(error);
-    res.status(error.status).json({ message: error.message });
-  }
-};
-
-const getReviews = async (req, res) => {
-  try {
-    const { productId } = req.params;
-    const reviews = await productsService.getReviews(productId);
+    // const { page, sort, category } = req.query;
+    // console.log(queryParams);
+    // console.log(Object.entries(queryParams));
+    const userId = req.user_id;
+    const productList = await productsService.getProductList(
+      userId
+      // page,
+      // sort,
+      // category
+    );
+    const totalProduct = await productsService.getTotalProduct();
     res.status(200).json({
       message: "querySuccess",
-      data: reviews,
+      data: productList,
+      productCount: totalProduct[0].product_count,
     });
   } catch (error) {
-    console.log(error);
-    res.status(error.status).json({message: error.message});
+    console.log("error", error);
+    res.status(error.status).json({ message: error.message });
   }
 };
 
 module.exports = {
-  createLike,
-  deleteLike,
-  getReviews,
+  getProduct,
 };
