+++ src/controllers/users.controller.js
@@ -1,23 +1,26 @@
 const { usersService } = require("../services");
 
 const signUp = async (req, res) => {
-    try {
-      const { password, email, name, phoneNumber } = req.body;
-      const complete = await usersService.signUp(name, email, password, phoneNumber)
+  try {
+    const { password, email, name, phoneNumber } = req.body;
+    const complete = await usersService.signUp(
+      name,
+      email,
+      password,
+      phoneNumber
+    );
 
-      return res.status(201).json({
-        message: "userCreated"
-      });
-    } catch (error) {
-      console.log(error);
-      return res.status(error.statusCode).json({
-        message: error.message,
-      });
-    }
-  };
+    return res.status(201).json({
+      message: "userCreated",
+    });
+  } catch (error) {
+    console.log(error);
+    return res.status(error.statusCode).json({
+      message: error.message,
+    });
+  }
+};
 
-
-  
 const signIn = async (req, res) => {
   try {
     const { email, password } = req.body;
@@ -34,9 +37,7 @@ const signIn = async (req, res) => {
   }
 };
 
-
 module.exports = {
   signIn,
   signUp,
 };
-
