+++ src/routes/products.router.js
@@ -1,12 +1,19 @@
 const express = require("express");
 
-const { productsController } = require("../controllers");
+const {
+  productsController,
+  likesController,
+  reviewsController,
+} = require("../controllers");
 
 const productsRouter = express.Router();
 
-productsRouter.post("/:productId/likes", productsController.createLike);
-productsRouter.delete("/:productId/likes", productsController.deleteLike);
-productsRouter.get("/:productId/reviews", productsController.getReviews);
+productsRouter.get("/", productsController.getProduct);
+
+productsRouter.post("/:productId/likes", likesController.createLike);
+productsRouter.delete("/:productId/likes", likesController.deleteLike);
+
+productsRouter.get("/:productId/reviews", reviewsController.getReviews);
 
 module.exports = {
   productsRouter,
