+++ src/services/index.js
@@ -1,7 +1,11 @@
 const usersService = require("./users.service");
 const productsService = require("./products.service");
+const likesService = require("./likes.service");
+const reviewsService = require("./reviews.service");
 
 module.exports = {
   usersService,
-  productsService
+  productsService,
+  likesService,
+  reviewsService,
 };
