+++ src/utils/checkEmptyValues.js
@@ -1,3 +1,5 @@
+const {throwError} = require("./throwError");
+
 function checkEmptyValues() {
   for (let i = 0; i < arguments.length; i++) {
     if (!arguments[i]) {
